# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CyberGame API",
        "contact": {}
    },
    "paths": {
...
        "/settingweb/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ตั้งค่าหน้าเว็บไซต์",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settingweb"
                ],
                "summary": "CreateSettingWeb",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SettingwebCreateBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/settingweb/detail/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลการตั้งค่า ด้วย id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settingweb"
                ],
                "summary": "GetSettingWebById",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessWithData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "model.SettingwebCreateBody": {
            "type": "object",
            "required": [
                "autowithdraw",
                "backgrondcolor",
                "depositFirst",
                "depositNext",
                "line",
                "opt",
                "otpRegister",
                "register",
                "tranWithdraw",
                "url",
                "userAuto",
                "withdraw"
            ],
            "properties": {
                "autowithdraw": {
                    "type": "string"
                },
                "backgrondcolor": {
                    "type": "string"
                },
                "depositFirst": {
                    "type": "string"
                },
                "depositNext": {
                    "type": "string"
                },
                "line": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "opt": {
                    "type": "string"
                },
                "otpRegister": {
                    "type": "string"
                },
                "register": {
                    "type": "string"
                },
                "tranWithdraw": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userAuto": {
                    "type": "string"
                },
                "withdraw": {
                    "type": "string"
                }
            }
        },
        "model.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.SuccessWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}